+++ src/main/java/capstone/restaurant/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package capstone.restaurant;
+package capstone.restaurant.config;
 
 import io.swagger.v3.oas.annotations.OpenAPIDefinition;
 import io.swagger.v3.oas.annotations.info.Info;
