+++ src/main/java/capstone/restaurant/dto/restaurant/RestaurantListResponse.java
@@ -1,4 +1,4 @@
-package capstone.restaurant;
+package capstone.restaurant.dto.restaurant;
 
 import lombok.Getter;
 import lombok.Setter;
